requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
click==7.1.2
Django==3.2
django-debug-toolbar==3.2.1
django-ninja==0.13.0
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.4.0
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8.2
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
gunicorn==20.1.0
mysqlclient==2.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
django-tempus-dominus==5.1.2.17
